Downgrade time dependency and remove HailgunTime
This is necessary because we don't want all the extra pinned packages in
stack.yaml, and I don't want to recreate the timeParseTimeSpecifier
function that is not exposed in the previous version, which necessitates
a removal of HailgunTime. This is hacky, but I need to upgrade our
other dependencies quickly.